requirements.txt
scikit-rf
pyserial
plotly
